@@ -32,89 +33,87 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
     "@heroicons/react": "^2.2.0",
-    "@mui/icons-material": "^7.2.0",
-    "@tailwindcss/cli": "^4.1.11",
-    "@tailwindcss/postcss": "^4.1.11",
-    "@types/jszip": "^3.4.0",
-    "@uiw/react-codemirror": "^4.24.0",
-    "@vercel/analytics": "^1.5.0",
-    "autoprefixer": "^10.4.21",
+    "@mui/icons-material": "^7.3.8",
+    "@uiw/react-codemirror": "^4.25.4",
+    "@vercel/analytics": "^1.6.1",
     "clsx": "^2.1.1",
-    "framer-motion": "^12.23.0",
-    "i18next": "^25.3.1",
-    "i18next-browser-languagedetector": "^8.2.0",
+    "framer-motion": "^12.34.0",
+    "i18next": "^25.8.7",
+    "i18next-browser-languagedetector": "^8.2.1",
     "jszip": "^3.10.1",
-    "lookatni-file-markers": "^1.1.0",
-    "lucide-react": "^0.525.0",
-    "next": "15.3.5",
-    "next-font": "^1.0.1",
+    "lookatni-file-markers": "^1.2.0",
+    "lucide-react": "^0.564.0",
+    "next": "^16.1.6",
+    "next-font": "^1.2.0",
     "next-themes": "^0.4.6",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-i18next": "^15.6.0",
-    "react-icons": "^4.6.0",
-    "react-intersection-observer": "^9.0.0",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "react-i18next": "^16.5.4",
+    "react-icons": "^5.5.0",
+    "react-intersection-observer": "^10.0.2",
     "react-toastify": "^11.0.5",
-    "resend": "^4.6.0",
-    "simple-icons": "^15.5.0",
-    "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.11"
+    "resend": "^6.9.2",
+    "simple-icons": "^16.8.0"
   },
   "devDependencies": {
     "@types/babel__core": "^7.20.5",
     "@types/babel__generator": "^7.27.0",
     "@types/babel__template": "^7.4.4",
-    "@types/babel__traverse": "^7.20.7",
+    "@types/babel__traverse": "^7.28.0",
     "@types/body-parser": "^1.19.6",
     "@types/bonjour": "^3.5.13",
     "@types/connect": "^3.4.38",
     "@types/connect-history-api-fallback": "^1.5.4",
-    "@types/eslint": "^9.6.1",
-    "@types/eslint-scope": "^8.3.0",
     "@types/estree": "^1.0.8",
-    "@types/express": "^4.17.23",
-    "@types/express-serve-static-core": "^5.0.6",
+    "@types/express": "^5.0.6",
+    "@types/express-serve-static-core": "^5.1.1",
     "@types/graceful-fs": "^4.1.9",
     "@types/html-minifier-terser": "^7.0.2",
     "@types/http-errors": "^2.0.5",
-    "@types/http-proxy": "^1.17.16",
+    "@types/http-proxy": "^1.17.17",
     "@types/istanbul-lib-coverage": "^2.0.6",
     "@types/istanbul-lib-report": "^3.0.3",
     "@types/istanbul-reports": "^3.0.4",
-    "@types/jsdom": "^21.1.7",
+    "@types/jsdom": "^27.0.0",
     "@types/json-schema": "^7.0.15",
-    "@types/json5": "^0.0.30",
-    "@types/mime": "^3.0.4",
-    "@types/node": "^20.19.4",
-    "@types/node-forge": "^1.3.12",
-    "@types/parse-json": "^4.0.2",
-    "@types/prettier": "^2.7.3",
+    "@types/node": "^25.2.3",
+    "@types/node-forge": "^1.3.14",
     "@types/q": "^1.5.8",
     "@types/qs": "^6.14.0",
     "@types/range-parser": "^1.2.7",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
     "@types/resolve": "^1.20.6",
     "@types/retry": "^0.12.5",
-    "@types/semver": "^7.7.0",
-    "@types/send": "^0.17.5",
+    "@types/semver": "^7.7.1",
+    "@types/send": "^0.17.6",
     "@types/serve-index": "^1.9.4",
-    "@types/serve-static": "^1.15.8",
+    "@types/serve-static": "^1.15.10",
     "@types/sockjs": "^0.3.36",
     "@types/stack-utils": "^2.0.3",
     "@types/tough-cookie": "^4.0.5",
     "@types/trusted-types": "^2.0.7",
     "@types/ws": "^8.18.1",
-    "@types/yargs": "^17.0.33",
+    "@types/yargs": "^17.0.35",
     "@types/yargs-parser": "^21.0.3",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.5",
+    
+    "@types/eslint": "^9.6.1",
+    "@types/eslint-scope": "^8.4.0",
+    "eslint": "^9.39.0",
+    "eslint-config-next": "16.1.6",
     "husky": "^9.1.7",
-    "jest": "^30.0.4",
-    "jest-environment-jsdom": "^30.0.4",
-    "lint-staged": "^16.1.2",
-    "multer": ">=2.0.2",
-    "serve": "^14.0.1",
+    "jest": "^30.2.0",
+    "jest-environment-jsdom": "^30.2.0",
+    "lint-staged": "^16.2.7",
+    "multer": "^2.0.2",
+    "serve": "^14.2.5",
+
+    "autoprefixer": "^10.4.24",
+    "tailwind-merge": "^3.4.0",
+    "tailwindcss": "^4.1.18",
+    "postcss": "^8.4.38",
+    "@tailwindcss/postcss": "^4.1.18",
+
     "typescript": "5.9.3"
   },
   "keywords": [
@@ -155,9 +154,9 @@
     ]
   },
   "resolutions": {
-    "next": "15.3.5",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "next": "^16.1.6",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4"
   },
   "prettier": {
     "singleQuote": true,
